+++ ReQuest/ui/gm/views.py
@@ -39,22 +39,12 @@ from ReQuest.utilities.supportFunctions import (
     replace_cached_data,
     escape_markdown,
     get_guild_member,
-    build_cache_key
+    build_cache_key,
+    check_role_hierarchy
 )
 
 logger = logging.getLogger(__name__)
 
-
-def check_role_hierarchy(guild: discord.Guild, role: discord.Role):
-    """Raises UserFeedbackError if the bot cannot manage the given role due to hierarchy."""
-    bot_top_role = guild.me.top_role
-    if role >= bot_top_role:
-        raise UserFeedbackError(
-            t(DEFAULT_LOCALE, 'gm-error-role-hierarchy', roleName=role.name, roleId=str(role.id)),
-            message_id='gm-error-role-hierarchy',
-            roleName=role.name,
-            roleId=str(role.id)
-        )
 
 
 class GMBaseView(MenuBaseView):
