+++ ReQuest/utilities/supportFunctions.py
@@ -40,6 +40,19 @@ class UserFeedbackError(Exception):
         return str(self)
 
 
+def check_role_hierarchy(guild: discord.Guild, role: discord.Role):
+    """Raises UserFeedbackError if the bot cannot manage the given role due to hierarchy."""
+    from ReQuest.utilities.localizer import t, DEFAULT_LOCALE
+    bot_top_role = guild.me.top_role
+    if role >= bot_top_role:
+        raise UserFeedbackError(
+            t(DEFAULT_LOCALE, 'gm-error-role-hierarchy', roleName=role.name, roleId=str(role.id)),
+            message_id='gm-error-role-hierarchy',
+            roleName=role.name,
+            roleId=str(role.id)
+        )
+
+
 def build_cache_key(database_name, identifier, collection_name):
     return f'{database_name}:{identifier}:{collection_name}'
